refactor: navigation query moved to layout
