+++ snapshot.gql
@@ -1,4 +1,4 @@
-### Type definitions saved at 2023-03-01T13:24:56.086Z ###
+### Type definitions saved at 2023-03-02T08:59:20.249Z ###
 
 enum RemoteFileFit {
   COVER
@@ -704,6 +704,7 @@ type StrapiNavigation implements Node @derivedTypes @dontInfer {
   external: Boolean
   items: [StrapiNavigationItems]
   highlight: Boolean
+  related: StrapiNavigationRelated
 }
 
 type StrapiNavigationItems @derivedTypes {
@@ -721,6 +722,20 @@ type StrapiNavigationItems @derivedTypes {
 }
 
 type StrapiNavigationItemsRelated {
+  id: Int
+  title: String
+  slug: String
+  createdAt: Date @dateformat
+  updatedAt: Date @dateformat
+  publishedAt: Date @dateformat
+  url_path_id: String
+  locale: String
+  _xcontentType: String @proxy(from: "__contentType")
+  navigationItemId: Int
+  _xtemplateName: String @proxy(from: "__templateName")
+}
+
+type StrapiNavigationRelated {
   id: Int
   title: String
   slug: String
@@ -816,9 +831,9 @@ type STRAPI__MEDIAFormatsThumbnail {
 type STRAPI__COMPONENT_SHARED_SEO implements Node @dontInfer {
   metaTitle: String
   metaDescription: String
-  strapi_id: Int
   metaImage: STRAPI__MEDIA @link(by: "id", from: "metaImage___NODE")
   metaSocial: [STRAPI__COMPONENT_SHARED_META_SOCIAL] @link(by: "id", from: "metaSocial___NODE")
+  strapi_id: Int
 }
 
 type STRAPI__COMPONENT_SHARED_META_SOCIAL implements Node @dontInfer {
