+++ src/partials/Header/Header.tsx
@@ -1,44 +1,21 @@
 import classNames from 'classnames';
-import { graphql, Link, useStaticQuery } from 'gatsby';
+import { Link } from 'gatsby';
 import React, { useState } from 'react';
 import { Menu } from '../../components/Menu';
-import { useLocalizedQuery } from '../../hooks/useLocalizedQuery';
 import { Hamburger } from '../Hamburger';
 import { Logo } from '../Logo';
 import { Socials } from '../Socials';
 import './Header.sass';
 
-enum MENU {
-  RESERVED_MENU = 'ReservedMenu',
-  MAIN_MENU = 'MainMenu',
-}
-
-export const Header = () => {
+export const Header = ({
+  reservedMenu,
+  mainMenu,
+}: {
+  reservedMenu: Queries.MainNavigationItemFragment[];
+  mainMenu: Queries.MainNavigationItemFragment[];
+}) => {
   const [mobileMenuOpen, setMobileMenuOpen] = useState<boolean>(false);
   const handleMobileMenu = () => setMobileMenuOpen(prev => !prev);
-
-  const query: Queries.MainNavigationQuery = useStaticQuery(graphql`
-    query MainNavigation {
-      allStrapiNavigation {
-        nodes {
-          ...MainNavigationItem
-        }
-      }
-    }
-  `);
-
-  const { localeNodes: menuNodes } = useLocalizedQuery<
-    Queries.MainNavigationItemFragment,
-    Queries.MainNavigationQuery
-  >({
-    query,
-    type: 'allStrapiNavigation',
-  });
-
-  const reservedMenu = menuNodes?.filter(
-    node => node?.key === MENU.RESERVED_MENU
-  );
-  const mainMenu = menuNodes?.filter(node => node?.key === MENU.MAIN_MENU);
 
   return (
     <header className={classNames('header', mobileMenuOpen && 'menu-is-open')}>
