+++ src/types.ts
@@ -1,0 +1,7 @@
+export namespace PagoPA {
+  export enum MENU {
+    RESERVED_MENU = 'ReservedMenu',
+    MAIN_MENU = 'MainMenu',
+    FOOTER_MENU = 'FooterMenu',
+  }
+}
